remove unnecessary js file
