+++ page/src/components/index.js
@@ -1,16 +1,10 @@
 export { default as SideBar } from './SideBar';
 export { default as Header } from './Header';
 export { default as PageInfo } from './PageInfo';
-export { default as AccountInfo } from './AccountInfo';
-export { default as ScoringInfo } from './ScoringInfo';
-export { default as AssignmentInfo } from './AssignmentInfo';
-export { default as PartInfo } from './PartInfo';
 export { default as BottomPopup } from './BottomPopup';
 export { default as AssignmentBox } from './AssignmentBox';
-export { default as PartBox } from './PartBox';
 export { default as Problem } from './Problem';
 export { default as TimePicker } from './TimePicker';
-export { default as SimpleModal } from './SetQuestion';
 export { default as UseRequest } from './UseRequest';
 export { default as UserAnswer } from './UserAnswer';
 export { default as Loading } from './Loading';
